+++ sandbox/sample/migrations/0001_initial.py
@@ -42,7 +42,7 @@ class Migration(migrations.Migration):
                     ),
                 ),
                 (
-                    "media",
+                    "mediafile",
                     models.FileField(
                         blank=True,
                         default=None,