+++ sandbox/sample/urls.py
@@ -1,12 +1,12 @@
 from django.urls import path
 
-from .views import ImageItemDetailView
+from .views import ImageItemDetailView, ImageItemFormView
 
 
 app_name = "sample"
 
 
 urlpatterns = [
-    path("", BlogIndexView.as_view(), name="index"),
+    path("form/", ImageItemFormView.as_view(), name="imageitem-form"),
     path("<int:imageitem_pk>/", ImageItemDetailView.as_view(), name="imageitem-detail"),
 ]