+++ smart_media/admin.py
@@ -4,20 +4,52 @@ from .modelfields import SmartMediaField
 from .widgets import ClearableFileInputButton
 
 
-class SmartModelAdmin(django_admin.ModelAdmin):
+class SmartAdminMixin:
     """
-    A model admin class to include the right widget definition for model field
-    ``modelfields.SmartMediaField``.
+    A class to mix with a Django admin class to the use right widget definition for
+    model field ``modelfields.SmartMediaField``.
 
-    This is required because ModelAdmin translate model field to widgets using a list
-    of harcoded relations between fields and widgets. It results field definition in
-    ``SmartMediaField.formfield`` is ignored and translate ``SmartMediaField`` to
-    ``AdminFileWidget`` instead of expected ``ClearableFileInputButton``.
+    .. Note::
+        This is required because ModelAdmin translate model fields to widgets using a
+        list of harcoded relations between fields and widgets.
 
-    So either your model admin inherits from ``SmartModelAdmin`` or simply copy its
-    ``SmartModelAdmin.formfield_overrides`` into it.
+        It results in field definitions from ``SmartMediaField.formfield`` to be
+        ignored and ``SmartMediaField`` using admin widget ``AdminFileWidget`` instead
+        of expected ``ClearableFileInputButton``.
 
-    It is only required with model forms in Django admin.
+        This mixin will be safe to use with specific model admins that does not allows
+        to inherit from ``SmartModelAdmin``.
+
+    Example:
+
+        Your model admin just have to inherit from this class in addition to the model
+        admin class: ::
+
+            from django.contrib import admin
+            from smart_media.admin import SmartAdminMixin
+            from myapp.models import MyModel
+
+            @admin.register(MyModel)
+            class MyModelAdmin(SmartAdminMixin, admin.ModelAdmin):
+                ...
+
+    If you can't use it, just copy the source of
+    ``SmartAdminMixin.formfield_overrides`` (see source) into your own model admin.
+    """
+    formfield_overrides = {
+        SmartMediaField: {
+            "widget": ClearableFileInputButton,
+        },
+    }
+
+
+class SmartModelAdmin(SmartAdminMixin, django_admin.ModelAdmin):
+    """
+    A model admin class to include ``SmartAdminMixin``.
+
+    It is only required with model forms in Django admin. If your model admin already
+    inherit from a specific admin class, you will prefer to use the
+    ``SmartAdminMixin``.
 
     Example:
 
@@ -30,12 +62,5 @@ class SmartModelAdmin(django_admin.ModelAdmin):
             @admin.register(MyModel)
             class MyModelAdmin(SmartModelAdmin):
                 ...
-
-    Or copy the source of ``SmartModelAdmin.formfield_overrides`` (see source) into
-    your own model admin.
     """
-    formfield_overrides = {
-        SmartMediaField: {
-            "widget": ClearableFileInputButton,
-        },
-    }
+    pass